+++ src/text-input/index.js
@@ -1,2 +1,3 @@
 // @flow
-export { TextInput } from "./text-input";
+export { TextInput } from "./v1/text-input";
+export { TextInputV2 } from "./v2/text-input";